Copied contents of ko.js to ko-fix.js
